+++ R/tar_quarto_files.R
@@ -6,9 +6,10 @@
 #'   of `quarto::quarto_inspect()` and returns what `tarchetypes` needs to
 #'   know about the current Quarto project or document.
 #' @return A named list of important file paths in a Quarto project or document:
-#'   * `sources`: source files with `tar_load()`/`tar_read()` target dependencies
-#'     in R code chunks.
-#'   * `output`: output files that will be generated during `quarto::quarto_render()`.
+#'   * `sources`: source files with `tar_load()`/`tar_read()`
+#'     target dependencies in R code chunks.
+#'   * `output`: output files that will be generated during
+#'     `quarto::quarto_render()`.
 #'   * `input`: pre-existing files required to render the project or document,
 #'     such as `_quarto.yml`.
 #' @param path Character of length 1, either the file path